requirements.txt
# ── Web framework ─────────────────────────────────────────
fastapi==0.128.8
uvicorn[standard]==0.39.0
starlette==0.49.3

# ── Database ──────────────────────────────────────────────
SQLAlchemy[asyncio]==2.0.48
asyncpg==0.31.0
psycopg2-binary==2.9.11
alembic==1.13.2
GeoAlchemy2==0.17.1

# ── Pydantic / validation ─────────────────────────────────
pydantic==2.12.5
pydantic[email]==2.12.5
pydantic-settings==2.5.2
pydantic_core==2.41.5
email-validator==2.2.0

# ── Auth ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── HTTP client ───────────────────────────────────────────
httpx==0.27.2
requests==2.32.5

# ── Spatial / H3 ─────────────────────────────────────────
h3==4.4.2
GeoAlchemy2==0.17.1
geopandas==1.0.1
shapely==2.0.7
pyproj==3.6.1
pyogrio==0.11.1

# ── ML / Data ─────────────────────────────────────────────
numpy==2.0.2
pandas==2.3.3
scikit-learn==1.6.1
scipy==1.13.1
xgboost==2.1.4
lightgbm==4.6.0
joblib==1.5.3
networkx==3.2.1
osmnx==2.0.7

# ── Queue / async ─────────────────────────────────────────
celery==5.6.2
redis==7.0.1
kombu==5.6.2
amqp==5.3.1
billiard==4.2.4
vine==5.1.0
websockets==15.0.1

# ── Utilities ─────────────────────────────────────────────
faker==30.1.0
python-dotenv==1.2.1
packaging==26.0
certifi==2026.2.25
charset-normalizer==3.4.5
idna==3.11
urllib3==2.6.3
six==1.17.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
tzdata==2025.3
tzlocal==5.3.1
anyio==4.12.1
async-timeout==5.0.1
exceptiongroup==1.3.1
h11==0.16.0
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2
prompt_toolkit==3.0.52
wcwidth==0.6.0